# openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "asystant-api Managed Keys API",
    "version": "0.3.0",
    "description": "Server-to-server API that gives client services (companies or products) short-lived, budget-limited OpenRouter keys for their users. Each client authenticates with its own key. Budgets are integer USD micros per tenant and subject; keys are created with the OpenRouter management API, sealed at rest, revoked when budgets drop and reconciled against confirmed usage. Maximum request body: 64 KiB.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local development only; replace with your HTTPS deployment."
    }
  ],
  "security": [
    {
      "clientKey": []
    }
  ],
  "paths": {
    "/v1/managed/credentials": {
      "post": {
        "operationId": "issueCredential",
        "summary": "Get the subject's OpenRouter key for today",
        "description": "Resolves the smallest remaining budget of the tenant and the subject, then creates one OpenRouter key per tenant, subject, bucket and UTC day through the OpenRouter management API. Repeated calls on the same day return the same sealed key without creating another one. Call it from the client backend after authenticating the subject; tenant and subject must come from that verified session.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Key for the subject; response is `Cache-Control: no-store`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credential"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, unknown field or out-of-range value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or revoked client key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Per-peer admission limit reached; retry after the `Retry-After` seconds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service or database unavailable, or the managed flow is not configured (`code`: `managed_configuration_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "402": {
            "description": "The tenant or subject budget is exhausted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "No budget is configured for this tenant or subject.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Issuance pending reconciliation (`code`: `managed_reconciliation_pending`) or the allocation changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/managed/tenants/{tenant}/budget": {
      "parameters": [
        {
          "name": "tenant",
          "in": "path",
          "required": true,
          "description": "Tenant identifier chosen by the client (for example a school or company id). Percent-encode reserved characters.",
          "schema": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        }
      ],
      "get": {
        "operationId": "getTenantBudget",
        "summary": "Policies, balances and key states of a tenant",
        "responses": {
          "200": {
            "description": "Only data owned by the authenticated client.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetOverview"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, unknown field or out-of-range value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or revoked client key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Per-peer admission limit reached; retry after the `Retry-After` seconds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service or database unavailable, or the managed flow is not configured (`code`: `managed_configuration_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "setTenantBudget",
        "summary": "Set the tenant ceiling and its OpenRouter workspace",
        "description": "The ceiling must cover the sum of its subject budgets. Lowering it queues every live key of the tenant for revocation.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantBudgetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stored policy; repeating the same value is a no-op.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Policy"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, unknown field or out-of-range value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or revoked client key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Per-peer admission limit reached; retry after the `Retry-After` seconds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service or database unavailable, or the managed flow is not configured (`code`: `managed_configuration_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The workspace is not assigned to this client, or the workspace id is nil.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Subject budgets exceed the ceiling, or the workspace differs from existing budgets.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/managed/tenants/{tenant}/subjects/{subject}/budget": {
      "parameters": [
        {
          "name": "tenant",
          "in": "path",
          "required": true,
          "description": "Tenant identifier chosen by the client (for example a school or company id). Percent-encode reserved characters.",
          "schema": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        {
          "name": "subject",
          "in": "path",
          "required": true,
          "description": "Subject identifier inside the tenant (for example a teacher or user id).",
          "schema": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        }
      ],
      "put": {
        "operationId": "setSubjectBudget",
        "summary": "Set a subject budget inside the tenant ceiling",
        "description": "Lowering it queues the subject's live key for revocation; set 0 to stop issuance.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectBudgetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stored policy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Policy"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, unknown field or out-of-range value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or revoked client key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Per-peer admission limit reached; retry after the `Retry-After` seconds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service or database unavailable, or the managed flow is not configured (`code`: `managed_configuration_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The tenant ceiling is not configured yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Subject budgets would exceed the tenant ceiling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/managed/tenants/{tenant}/leases/{lease}/recovery": {
      "parameters": [
        {
          "name": "tenant",
          "in": "path",
          "required": true,
          "description": "Tenant identifier chosen by the client (for example a school or company id). Percent-encode reserved characters.",
          "schema": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        {
          "name": "lease",
          "in": "path",
          "required": true,
          "description": "Lease id from the budget overview.",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "authorizeRecovery",
        "summary": "Replace an uncertain key issuance",
        "description": "For a current lease in `uncertain` or `revocation_pending`, reserves a new allocation for today and queues the previous one for revocation. The previous reservation is never released. Repeating the same authorization returns the stored recovery.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authorized replacement.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recovery"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, unknown field or out-of-range value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or revoked client key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Per-peer admission limit reached; retry after the `Retry-After` seconds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service or database unavailable, or the managed flow is not configured (`code`: `managed_configuration_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "402": {
            "description": "The budget cannot cover the new allocation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The current policy does not allow the replacement.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Lease not found for this client and tenant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "The lease is not pending reconciliation, or another recovery or allocation exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "clientKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Client key issued by asystant-api (`ask_` + 64 hex characters). Keep it in the client backend's secret manager; never send it to a browser or app."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "enum": [
              "managed_configuration_required",
              "managed_creation_rejected",
              "managed_reconciliation_pending"
            ],
            "description": "Present on managed-key failures."
          }
        }
      },
...
      "UsageBucket": {
        "type": "string",
        "enum": [
          "daily",
          "migration"
        ],
        "description": "`daily` resets every UTC day; `migration` is a lifetime budget."
      },
      "CredentialRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "tenant",
          "subject",
          "bucket"
        ],
        "properties": {
          "tenant": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "subject": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "bucket": {
            "$ref": "#/components/schemas/UsageBucket"
          }
        }
      },
      "Credential": {
        "type": "object",
        "required": [
          "tenant",
          "subject",
          "bucket",
          "expires_at",
          "refresh_after",
          "allowed_models",
          "api_key"
        ],
        "properties": {
          "tenant": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "bucket": {
            "$ref": "#/components/schemas/UsageBucket"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "End of the current UTC day."
          },
          "refresh_after": {
            "type": "string",
            "format": "date-time",
            "description": "Ask again after this instant; the same key is returned until it changes."
          },
          "allowed_models": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "api_key": {
            "type": "string",
            "description": "OpenRouter key limited to the allocation. Deliver it only to the authenticated subject; never store or log it."
          }
        }
      },
      "TenantBudgetRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "workspace_id",
          "bucket",
          "limit_usd_micros",
          "actor"
        ],
        "properties": {
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "description": "OpenRouter workspace where the tenant's keys are created. It must be one of the workspaces assigned to the client when it was created, and cannot change once budgets exist."
          },
          "bucket": {
            "$ref": "#/components/schemas/UsageBucket"
          },
          "limit_usd_micros": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Integer USD micros: 1,000,000 = USD 1."
          },
          "actor": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Who made the change in the client, for the audit trail."
          }
        }
      },
      "SubjectBudgetRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "bucket",
          "limit_usd_micros",
          "actor"
        ],
        "properties": {
          "bucket": {
            "$ref": "#/components/schemas/UsageBucket"
          },
          "limit_usd_micros": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Integer USD micros: 1,000,000 = USD 1."
          },
          "actor": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        }
      },
      "Policy": {
        "type": "object",
        "required": [
          "id",
          "client_id",
          "tenant",
          "owner_kind",
          "owner_id",
          "workspace_id",
          "bucket",
          "limit_usd_micros",
          "updated_by",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "client_id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant": {
            "type": "string"
          },
          "owner_kind": {
            "type": "string",
            "enum": [
              "tenant",
              "subject"
            ]
          },
          "owner_id": {
            "type": "string"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid"
          },
          "bucket": {
            "$ref": "#/components/schemas/UsageBucket"
          },
          "limit_usd_micros": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Integer USD micros: 1,000,000 = USD 1."
          },
          "updated_by": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BudgetAccount": {
        "type": "object",
        "required": [
          "id",
          "tenant",
          "owner_kind",
          "owner_id",
          "workspace_id",
          "bucket",
          "period_key",
          "limit_usd_micros",
          "spent_usd_micros",
          "reserved_usd_micros"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "client_id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant": {
            "type": "string"
          },
          "owner_kind": {
            "type": "string",
            "enum": [
              "tenant",
              "subject"
            ]
          },
          "owner_id": {
            "type": "string"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid"
          },
          "bucket": {
            "$ref": "#/components/schemas/UsageBucket"
          },
          "period_key": {
            "type": "string",
            "description": "UTC date for `daily`, `lifetime` for `migration`."
          },
          "limit_usd_micros": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Integer USD micros: 1,000,000 = USD 1."
          },
          "spent_usd_micros": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Integer USD micros: 1,000,000 = USD 1."
          },
          "reserved_usd_micros": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Held by issued keys until OpenRouter confirms their usage; never released by expiry alone."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CredentialStatus": {
        "type": "object",
        "required": [
          "id",
          "subject",
          "is_current",
          "can_authorize_recovery",
          "bucket",
          "status",
          "limit_usd_micros",
          "expires_at",
          "accounted_usage_usd_micros"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subject": {
            "type": "string"
          },
          "is_current": {
            "type": "boolean"
          },
          "can_authorize_recovery": {
            "type": "boolean"
          },
          "bucket": {
            "$ref": "#/components/schemas/UsageBucket"
          },
          "status": {
            "type": "string",
            "enum": [
              "reserved",
              "provisioning",
              "uncertain",
              "issued",
              "revocation_pending",
              "revoked"
            ]
          },
          "limit_usd_micros": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Integer USD micros: 1,000,000 = USD 1."
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "accounted_usage_usd_micros": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Integer USD micros: 1,000,000 = USD 1."
          },
          "usage_checked_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "BudgetOverview": {
        "type": "object",
        "required": [
          "policies",
          "accounts",
          "credentials"
        ],
        "properties": {
          "policies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Policy"
            }
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BudgetAccount"
            }
          },
          "credentials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CredentialStatus"
            }
          }
        }
      },
      "RecoveryRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "limit_usd_micros",
          "reason",
          "acknowledge_pending_reserve",
          "actor"
        ],
        "properties": {
          "limit_usd_micros": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "Integer USD micros: 1,000,000 = USD 1."
          },
          "reason": {
            "type": "string",
            "minLength": 10,
            "maxLength": 500
          },
          "acknowledge_pending_reserve": {
            "type": "boolean",
            "const": true,
            "description": "The previous reservation stays held until OpenRouter confirms its usage."
          },
          "actor": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        }
      },
      "Recovery": {
        "type": "object",
        "required": [
          "id",
          "client_id",
          "tenant",
          "previous_lease_id",
          "replacement_lease_id",
          "actor",
          "limit_usd_micros",
          "reason",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "client_id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant": {
            "type": "string"
          },
          "previous_lease_id": {
            "type": "string",
            "format": "uuid"
          },
          "replacement_lease_id": {
            "type": "string",
            "format": "uuid"
          },
          "actor": {
            "type": "string"
          },
          "limit_usd_micros": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Integer USD micros: 1,000,000 = USD 1."
          },
          "reason": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}